gh-10: re-organize mongo-filter and realted logic
