+++ lib/mongo-filter/is-mongo-date.js
