+++ lib/mongo-filter/is-mongo-null.js
