+++ lib/mongo-filter/is-mongo-number.js
@@ -1,4 +1,4 @@
-const isString = require('./is-string')
+const isString = require('../helpers/is-string')
 
 // TODO doc and explain this 
 const isMongoNumber = (value) => {
