+++ lib/mongo-filter/parse-mongo-filter.js
@@ -1,37 +1,11 @@
 const qs = require('qs')
-const determineFilterStyle = require('./helpers/determine-filter-style')
-const FilterStyle = require('./enums/filter-style')
-const MongoOperator = require('./enums/mongo-operator')
-const SqlOperator = require('./enums/sql-operator')
-const isMongoNull = require('./mongo-filter/is-mongo-null')
-const areMongoTypesTheSame = require('./mongo-filter/are-mongo-types-the-same')
-const MongoValueType = require('./enums/mongo-value-type')
+const areMongoTypesTheSame = require('./are-mongo-types-the-same')
+const isMongoNull = require('./is-mongo-null')
+const MongoOperator = require('../enums/mongo-operator')
+const SqlOperator = require('../enums/sql-operator')
+const MongoValueType = require('../enums/mongo-value-type')
 
-const parseFilter = (queryString) => {
-  let results
-  const errors = []
-  let style
-
-  try {
-    style = determineFilterStyle(queryString)
-  } catch (e) {
-    errors.push(e)
-  }
-
-  if (style) {
-    const parserMap = {
-      [FilterStyle.MONGO_DB]: parseMongoDBFilter,
-      [FilterStyle.IBM]: parseIBMFIlter,
-    }
-    const styleParserReturn = parserMap[style](queryString)
-    results = styleParserReturn.results
-    errors.push(...styleParserReturn.errors)
-  }
-
-  return { results, errors }
-}
-
-const parseMongoDBFilter = (querystring) => {
+const parseMongoFilter = (querystring) => {
   const errors = []
   let results
 
@@ -196,11 +170,6 @@ const parseMongoDBFilter = (querystring) => {
   return { results, errors }
 }
 
-const parseIBMFIlter = (querystring) => {
-  // todo
-}
-
 module.exports = {
-  parseFilter,
-  parseMongoDBFilter
+  parseMongoFilter
 }
