+++ test/mongo-filter/are-mongo-types-the-same.test.js
@@ -1,5 +1,5 @@
 const MongoValueType = require('../../lib/enums/mongo-value-type')
-const areMongoTypesTheSame = require('../../lib/helpers/are-mongo-types-the-same')
+const areMongoTypesTheSame = require('../../lib/mongo-filter/are-mongo-types-the-same')
 
 describe('areMongoTypesTheSame', () => {
     /**
